Feat: Add in an option to use the ajv-formats plugin
Hi vanekj
We have been working on a Workflow that needed an AJV validator to help validate our Yaml files. When searching for AJV actions which could do the job, we stumbled upon yours, which we liked. This action was easy to understand, and we thought it could be the base of what we needed.

With these Actions, we have added a few needed features to make it work within our project and could benefit others looking for the same thing. We have added in:

An option to use ajv-format plugin. Great for when you use formats not supported by AJV alone.
An option to allow matching properties. We needed this for our projects because we used "properties" and "patternProperties".
An option to show validation results in a workflow summary. This is for quick analysis when you want to avoid looking at logs.
We hope you like the changes we have made.